Move format loading/dispatch into core.io; vcf.py is now just the CLI
Addresses a layering smell: the format dispatcher (read_variants) lived in
preprocessing/vcf.py, so even a plain-TSV input had to route through the VCF
module to reach the TSV loader. Loading variants in any format is an IO-layer
concern and belongs next to load_variants.

- core/io.py now owns load_variants_vcf, _is_symbolic_alt, and read_variants
  (the format-neutral dispatcher) alongside load_variants/VARIANT_SCHEMA. VCF
  parsing is pure-Python (stdlib gzip), so it keeps core TF-free and
  depends-on-nothing-internal.
- preprocessing/vcf.py is reduced to the thin converter CLI, calling
  io_utils.read_variants. validate.py calls io_utils.read_variants directly (no
  longer imports from the vcf module).
- tests/test_vcf.py targets core.io; docs/vcf.md + CLAUDE.md (conventions +
  module map) updated to point at core.io.read_variants.
